pyproject.toml
[tool.poetry]
name = "lunit_technical_case"
version = "0.1.0"
description = "ML technical project for Lunit"
authors = ["Paul Chauvin"]

[tool.poetry.dependencies]
python = "3.8"
torch = "^1.9.1"
torchvision = "^0.10.1"
mlflow = "^1.20.2"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
pydantic = "^1.8.2"
pyyaml = "^5.4.1"
Pillow = "^8.3.2"
numpy = "^1.21"
flask = "^3.0.0"
scikit-learn = "^1.3.1"
setuptools = "^68.2.2"
python-multipart = "^0.0.6"
transformers = "^4.34.1"
transformer-implementations = "^0.0.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"